pyproject.toml
[project]
name = "DataLens"
version = "0.1.0"
description = "A GenAI powered data analysis and forecasting tool built with Google Gemini and Streamlit"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-google-genai>=4.2.1",
    "pandas>=2.0.0,<3",
    "plotly>=6.6.0",
    "prophet>=1.3.0",
    "streamlit>=1.54.0",
    "python-dotenv>=1.0.0",
]
